+++ src/index.js
@@ -1,52 +1,13 @@
-import "./style.css";
+import './style.css';
+import * as add from './addLeader.js';
+
 /* eslint-disable max-classes-per-file */
+
 const list = document.getElementById('list');
-const leaderName = document.getElementById('name');
-const leaderScore = document.getElementById('score');
 const addButton = document.querySelector('.buttonClass');
 
-class Leader {
-  constructor(name, score) {
-    this.name = name;
-    this.score = score;
-  }
-}
-
-class UseLeader {
-  static createLeader() {
-    return new Leader(leaderName.value, leaderScore.value);
-  }
-
-  static saveLeader(newLeader) {
-    const Leaders = JSON.parse(localStorage.getItem('Leaders'));
-    if (Leaders === null) {
-      localStorage.setItem('Leaders', JSON.stringify([]));
-    } else {
-      Leaders.push(newLeader);
-      localStorage.setItem('Leaders', JSON.stringify(Leaders)); //
-    }
-  }
-
-  static findLeaders() {
-    return JSON.parse(localStorage.getItem('Leaders'));
-  }
-
-  static displayLeaders() {
-    const reloadLeaders = UseLeader.findLeaders() || [];
-    list.innerHTML = '';
-    reloadLeaders.forEach((aLeader) => {
-      const Leader = document.createElement('li');
-      Leader.innerHTML = `<p>${aLeader.name}: ${aLeader.score}</p>`;
-      const br = document.createElement('br');
-      list.appendChild(Leader);
-      list.appendChild(br);
-      
-    });
-  }
-}
-
 addButton.addEventListener('click', () => {
-  const newLeader = UseLeader.createLeader();
+  const newLeader = add.UseLeader.createLeader();
   UseLeader.saveLeader(newLeader);
   UseLeader.displayLeaders();
   const Leaders = UseLeader.findLeaders();
@@ -55,11 +16,9 @@ addButton.addEventListener('click', () => {
     const Leader = document.createElement('li');
     Leader.innerHTML = `<p>${aLeader.name}</p>
         <p>${aLeader.score} </p>`;
-    deleteBtn.id = aLeader.name;
     const br = document.createElement('br');
     list.appendChild(Leader);
     list.appendChild(br);
-    
     UseLeader.saveLeader(aLeader);
   }
 });
