+++ src/jshintrc.json
@@ -1,0 +1,14 @@
+{
+ "camelcase" : true,
+ "indent": 2,
+ "undef": true,
+ "quotmark": false,
+ "maxlen": 80,
+ "trailing": true,
+ "curly": true,
+ "strict": false,
+ "browser": true,
+ "devel": true,
+ "jquery": true,
+ "esversion": 6
+}
