+++ src/singleLeader.js
@@ -1,0 +1,6 @@
+export class Leader {
+  constructor(name, score) {
+    this.name = name;
+    this.score = score;
+  }
+}
