+++ src/style.css
@@ -1,5 +1,5 @@
 body {
-  font-family:'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+  font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
 }
 
 header {
@@ -41,7 +41,7 @@ form {
   border: solid rgb(202, 202, 202);
   border-radius: 5px; */
   margin: 20px;
-  padding: 10px;  
+  padding: 10px;
   width: 400px;
   text-align: right;
 }
@@ -55,7 +55,7 @@ form {
   border: solid rgb(202, 202, 202);
   border-radius: 5px; */
   margin: 20px;
-  padding: 10px;  
+  padding: 10px; 
   width: 400px;
 }
 
@@ -64,7 +64,7 @@ form {
   justify-content: space-between;
   height: 50px;
   align-items: center;
-  padding: 10px; 
+  padding: 10px;
 }
 
 .scoreList {
@@ -73,6 +73,5 @@ form {
   padding: 20px;
   margin: 10px;
   box-shadow: 0 0 10px gray;
-  border: solid rgb(202, 202, 202);
   border-radius: 5px;
 }
